backend/requirements.txt
# FastAPI 프레임워크와 서버
fastapi==0.115.5          # FastAPI 프레임워크 (메인!)
uvicorn[standard]==0.32.1  # ASGI 서버 (FastAPI를 실행하는 엔진)

# OpenAI API
openai==1.54.5            # OpenAI Python SDK

# 환경 변수 관리
python-dotenv==1.0.1      # .env 파일에서 API 키 읽기

# 데이터 검증
pydantic==2.10.3          # 데이터 검증 (FastAPI가 자동으로 사용)

# CORS (React 연결용)
# fastapi에 포함되어 있어서 별도 설치 불필요

# ============================================
# RAG (Retrieval-Augmented Generation) 관련
# ============================================

# 벡터 데이터베이스 (최신 버전, Windows 호환성 개선)
chromadb>=0.5.0           # 임베딩 벡터 저장 및 검색

# 임베딩 모델
sentence-transformers>=2.2.0  # BGE-M3-KO 모델 로딩

# 딥러닝 프레임워크
torch>=2.0.0              # PyTorch (CPU 버전, GPU 필요시 별도 설치)

# 문서 처리
pypdf>=3.0.0              # PDF 파일 읽기
python-docx>=1.0.0        # DOCX 파일 읽기
tiktoken>=0.5.0           # 토큰 카운팅 (OpenAI 토큰 계산)

# LangChain (최신 버전)
langchain>=0.1.0          # RAG 체인 구성
langchain-openai>=0.0.2   # LangChain OpenAI 통합
langchain-community>=0.0.10  # 커뮤니티 통합

# ============================================
# MCP (Model Context Protocol) 클라이언트
# ============================================
fastmcp>=2.0.0            # FastMCP 라이브러리 (Tavily MCP 연동)
